+++ ios/TourGraph/TourGraph/TourGraphWidgets/TourGraphWidgetsBundle.swift
@@ -1,0 +1,11 @@
+import WidgetKit
+import SwiftUI
+
+@main
+struct TourGraphWidgetsBundle: WidgetBundle {
+    var body: some Widget {
+        RightNowWidget()
+        RandomTourWidget()
+        LockScreenWidget()
+    }
+}